KnowledgeBase/requirements.txt
transformers
torch
psycopg2-binary
numpy
sqlalchemy
pgvector
